+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Version 0.4.2-dev (changes since 0.4.2 go here)
 
 ### Features
-
+* Added `skbio.sequence.replace` for replacing items in a `Sequence` with a given character. ([#1222](https://github.com/biocore/scikit-bio/issues/1222))
 ### Backward-incompatible changes [stable]
 
 ### Backward-incompatible changes [experimental]